Version 4.0
Changed the :PosttoTwitter command so that you now just type your Twitter in the command-line. So when you do :PosttoTwitter it will ask you for your twitter you type it in and send it off; which is better because there&#039s no need to start a new buffer. To Twitter the current buffer now do :BPosttoTwitter, and to Twitter the current line do the :CPosttoTwitter.